add soft background to the main content column
Uses the CSS3 multiple backgrounds feature - this means that Firefox 3.5 and older (and of course all IE's) won't show the background.
Firefox 3.6+, Opera 10.53+ as well as recent versions of Konqueror, Chrome, and Safari should show it just fine though.
In order to make it work with older browsers, several (probably 8) new div's would have to be added, which I figured was not worth it
for a non-essential feature like this.