Split CSS into several files; serve a dynamically generated, cached, combined + minimized version to the browser
The css markup has been cleaned up, reorganized and logically separated into multiple files which are now located in the newly created css/theme4/ directory.
The index.html will however serve a single combined CSS file to the browser, which is dynamically re-generated whenever a file in the css/theme4/ directory has changed.

Two new functions were added to the inc/functions.php file to achieve this, and the third-party CSS Compressor that was included in the previous commit (inc/css_compressor.php) was modified to meet these needs. (In it's current (comment-heavy) form the combined css file size gets compressed by about 46% thanks to the compressor).

Note that the Apache process needs write permission to the dynamic/ folder in order for all of this to work.